refactor: change UnreferencedVarDeclWalker.cpp to UnreferencedVarDeclWalker.hpp to include hpp file not cpp
